follow hyrax's naming convention for actors
