+++ app/actors/hyrax/actors/image_actor.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 module Hyrax
   module Actors
-    class ImageActor < ::Spot::BaseActor
+    class ImageActor < ::Spot::Actors::BaseActor
     end
   end
 end
