+++ app/actors/hyrax/actors/publication_actor.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 module Hyrax
   module Actors
-    class PublicationActor < ::Spot::BaseActor
+    class PublicationActor < ::Spot::Actors::BaseActor
       private
 
         # Overrides the +Hyrax::Actors::BaseActor+ method so that we can apply a +date_available+
